fix: change format of dbsync testool files
