+++ src/common/dbsync/testtool/cmdArgsHelper.hpp
@@ -1,16 +1,4 @@
-/*
- * Wazuh DBSYNC
- * Copyright (C) 2015, Wazuh Inc.
- * July 02, 2020.
- *
- * This program is free software; you can redistribute it
- * and/or modify it under the terms of the GNU General Public
- * License (version 2) as published by the FSF - Free Software
- * Foundation.
- */
-
-#ifndef _CMD_LINE_ARGS_HELPER_H_
-#define _CMD_LINE_ARGS_HELPER_H_
+#pragma once
 
 #include <iostream>
 #include <sstream>
@@ -91,5 +79,3 @@ private:
     const std::string m_outputFolder;
     const std::vector<std::string> m_actions;
 };
-
-#endif // _CMD_LINE_ARGS_HELPER_H_
