+++ src/common/dbsync/testtool/testContext.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "dbsync.h"
+#include "dbsync.hpp"
+
+struct TestContext
+{
+    DBSYNC_HANDLE handle;
+    TXN_HANDLE txnContext;
+    size_t currentId;
+    std::string outputPath;
+};
